refactoring: fsd 파일 구조에 맞게 파일 구조 재작업
