+++ src/pages/auth/index.ts
@@ -1,1 +1,2 @@
-export { default as SignupPage } from "./ui/SignupPage"; export { default as SharedPage } from "./ui/SharedPage";
+export { default as SignupPage } from "./ui/SignupPage"; 
+export { default as SharedPage } from "./ui/SharedPage";
