+++ src/widgets/event/eventCard/index.tsx
