--- src/widgets/eventCard/index.ts
@@ -1,1 +1,0 @@
-export { default as EventCard } from "./ui/EventCard";
