+++ src/widgets/group/groupCard/index.tsx
