--- src/widgets/groupCard/index.ts
@@ -1,1 +1,0 @@
-export { default as GroupCard } from "./ui/GroupCard";
