+++ src/widgets/home/guideCard/index.tsx
