+++ src/widgets/home/homeBanner/index.tsx
